Montse
Se implementó la modificación de un Turno, la eliminación.

+++ CIDFares.Spa.WFApplication/Forms/Catalogos/FrmTurnoPrincipal.cs
@@ -11,6 +11,7 @@ using CIDFares.Library.Controls.CIDWait.Code;
 using CIDFares.Spa.Business.ValueObjects;
 using CIDFares.Spa.DataAccess.Contracts.Entities;
 using CIDFares.Spa.WFApplication.Session;
+using System.Threading.Tasks;
 
 namespace CIDFares.Spa.WFApplication.Forms.Catalogos
 {
@@ -40,10 +41,10 @@ namespace CIDFares.Spa.WFApplication.Forms.Catalogos
         {
             try
             {
-                CIDWait.Show(async () =>
-                {
+                CIDWait.Show(async () => {
                     await Model.GetListaTurno();
                 }, "Espere");
+                
                 IniciarBinding();
             }
             catch (Exception ex)
@@ -56,8 +57,9 @@ namespace CIDFares.Spa.WFApplication.Forms.Catalogos
         {
             try
             {
-                FrmTurno Turno = new FrmTurno();
+                FrmTurno Turno = new FrmTurno(0);
                 Turno.ShowDialog();
+                GetDataAsync();
             }
             catch (Exception ex)
             {
@@ -70,13 +72,34 @@ namespace CIDFares.Spa.WFApplication.Forms.Catalogos
         {
             try
             {
-                FrmTurno Turno = new FrmTurno();
-                Turno.ShowDialog();
+                Turno item = ObtenerSeleccionado();
+                if (item != null)
+                {
+                    FrmTurno frmTurno = new FrmTurno(item.IdTurno);
+                    frmTurno.ShowDialog();
+                    frmTurno.Dispose();
+                    GetDataAsync();
+                }
+                else
+                    CIDMessageBox.ShowAlert(Constants.Messages.SystemName, Constants.Messages.GridSelectMessage, TypeMessage.informacion);
             }
             catch (Exception ex)
             {
-                ErrorLogHelper.AddExcFileTxt(ex, "FrmEmpleadoGrid ~ BtnNuevo_Click(object sender, EventArgs e)");
-                CIDMessageBox.ShowAlert(Messages.SystemName, Messages.ErrorMessage, TypeMessage.error);
+                throw ex;
+            }
+        }
+
+        private void GetDataAsync()
+        {
+            try
+            {
+                CIDWait.Show(async () => {
+                    await Model.GetListaTurno();
+                }, "Espere");
+            }
+            catch (Exception ex)
+            {
+                throw ex;
             }
         }
 
@@ -85,22 +108,31 @@ namespace CIDFares.Spa.WFApplication.Forms.Catalogos
             try
             {
                 var item = ObtenerSeleccionado();
+                Model.IdTurno = item.IdTurno;
                 if (item != null)
                 {
-                    if (CIDMessageBox.ShowAlertRequest(Messages.SystemName, Messages.ConfirmDeleteMessage) == DialogResult.OK)
+                    if (Model.IdTurno == 1 || Model.IdTurno == 2 || Model.IdTurno == 3)
                     {
-                        Model.IdTurno = item.IdTurno;
-                        int result = 0;
-                        result = await Model.Delete(CurrentSession.IdCuentaUsuario);
-                        if (result > 0)
-                        {
-                            CIDMessageBox.ShowAlert(Messages.SystemName, Messages.SuccessDeleteMessage, TypeMessage.correcto);
-                            await Model.GetListaTurno();
-                            //await Model.GetAllAsync();
-                        }
-                        else
-                            CIDMessageBox.ShowAlert(Messages.SystemName, Messages.ErrorDeleteMessage, TypeMessage.error);
+                        CIDMessageBox.ShowAlert(Messages.SystemName, Messages.NoEliminarTurno, TypeMessage.error);
                     }
+                    else
+                    {
+                        if (CIDMessageBox.ShowAlertRequest(Messages.SystemName, Messages.ConfirmDeleteMessage) == DialogResult.OK)
+                        {
+                            //Model.IdTurno = item.IdTurno;
+                            int result = 0;
+                            result = await Model.Delete(CurrentSession.IdCuentaUsuario);
+                            if (result > 0)
+                            {
+                                CIDMessageBox.ShowAlert(Messages.SystemName, Messages.SuccessDeleteMessage, TypeMessage.correcto);
+                                await Model.GetListaTurno();
+                                //await Model.GetAllAsync();
+                            }
+                            else
+                                CIDMessageBox.ShowAlert(Messages.SystemName, Messages.ErrorDeleteMessage, TypeMessage.error);
+                        }
+                    }
+                    //NoEliminarTurno
                 }
                 else
                     CIDMessageBox.ShowAlert(Messages.SystemName, Messages.GridSelectMessage, TypeMessage.informacion);
